optimised metrics2,3,4, added metrics5, revamp into system mode
